chore: league detail 컴포넌트화
